# rs-frontend/src/app/model/roomRequestDto.ts
/**
 * RoomEquipment API
 * This API is used to provide availability to customize rooms and their equipment. <br><br> Only users with given authorities can use it: <br><ul> <li>ADMIN</li> <li>SUPER ADMIN</li> </ul>
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

/**
 * Request object for adding oar modifying room
 */
export interface RoomRequestDto { 
    /**
     * Unique number of room
     */
    roomNumber?: number;
    /**
     * Room description
     */
    description?: string;
    roomStatus?: RoomRequestDto.RoomStatusEnum;
    equipmentItemsId?: Array<number>;
}
export namespace RoomRequestDto {
    export type RoomStatusEnum = 'NORMAL' | 'ONLY_TEACHER' | 'ONLY_PRINCIPAL';
    export const RoomStatusEnum = {
        NORMAL: 'NORMAL' as RoomStatusEnum,
        ONLYTEACHER: 'ONLY_TEACHER' as RoomStatusEnum,
        ONLYPRINCIPAL: 'ONLY_PRINCIPAL' as RoomStatusEnum
    };
}